首页商品信息展示

+++ Models/Coupon.cs
@@ -19,9 +19,10 @@ namespace InternetMall.Models
         public decimal? Threshold { get; set; }
         public decimal? Discount1 { get; set; }
         public decimal? Discount2 { get; set; }
-        public int ? Category { get; set; }
+        public short? Category { get; set; }
         public string ShopId { get; set; }
         public string CommodityId { get; set; }
+
         public virtual Commodity Commodity { get; set; }
         public virtual Shop Shop { get; set; }
         public virtual ICollection<BuyerCoupon> BuyerCoupons { get; set; }
